+++ pakyow-core/lib/pakyow/behavior/verification.rb
@@ -1,5 +1,7 @@
 # frozen_string_literal: true
 
+require "forwardable"
+
 require "pakyow/support/class_state"
 require "pakyow/support/extension"
 
@@ -10,27 +12,64 @@ module Pakyow
     module Verification
       extend Support::Extension
 
-      def verify(values = nil, &block)
-        unless values
-          if self.class.object_name_to_verify.nil?
+      def verify(name_or_values = name_or_values_omitted = true, values_or_name = values_or_name_omitted = true, &block)
+        name, values = :default, nil
+        if name_or_values_omitted
+          # intentionally empty
+        elsif name_or_values.is_a?(Symbol)
+          name = name_or_values
+          unless values_or_name_omitted
+            values = values_or_name
+          end
+        else
+          values = name_or_values
+        end
+
+        if values.nil?
+          if self.class.__verifiable_object_name.nil?
             raise "Expected values to be passed"
           else
-            values = public_send(self.class.object_name_to_verify)
+            values = public_send(self.class.__verifiable_object_name)
           end
         end
 
-        Pakyow::Verifier.new(&block).call!(values, context: self)
+        verifier = if block_given?
+          Pakyow::Verifier.new(&block)
+        else
+          self.class.__verifiers[name]
+        end
+
+        verifier&.call!(values, context: self)
       end
 
       apply_extension do
         extend Support::ClassState
-        class_state :object_name_to_verify, inheritable: true
+        class_state :__verifiable_object_name, inheritable: true
+        class_state :__verifiers, default: {}, inheritable: true
       end
 
       class_methods do
-        def verifies(object)
-          @object_name_to_verify = object
+        def verifies(name)
+          @__verifiable_object_name = name
         end
+
+        def verify(name = :default, &block)
+          name = name.to_sym
+          unless verifier = __verifiers[name]
+            verifier = if block_given?
+              Pakyow::Verifier.new(&block)
+            else
+              Pakyow::Verifier.new
+            end
+
+            __verifiers[name] = verifier
+          end
+
+          verifier
+        end
+
+        extend Forwardable
+        def_delegators :verify, :required, :optional
       end
     end
   end