+++ pakyow-core/spec/features/verification/dynamic_instance_spec.rb
@@ -1,23 +1,27 @@
 require "pakyow/behavior/verification"
 
-RSpec.describe Pakyow::Behavior::Verification do
-  include Pakyow::Behavior::Verification
+RSpec.describe "dynamic instance-level verification" do
+  let(:verifiable) {
+    Class.new do
+      include Pakyow::Behavior::Verification
+    end
+  }
 
-  before do
-    self.class.instance_variable_set(:@object_name_to_verify, nil)
-  end
+  let(:target) {
+    verifiable.new
+  }
 
-  context "object to verify is not defined" do
+  context "verifiable object is not defined" do
     context "values are passed" do
-      let :values do
+      let(:values) {
         {
           foo: "foo",
           bar: "bar",
         }
-      end
+      }
 
       before do
-        verify values do
+        target.verify values do
           optional :foo
         end
       end
@@ -30,7 +34,7 @@ RSpec.describe Pakyow::Behavior::Verification do
     context "values are not passed" do
       it "raises an error" do
         expect {
-          verify do
+          target.verify do
             optional :foo
           end
         }.to raise_error(RuntimeError) do |error|
@@ -40,28 +44,35 @@ RSpec.describe Pakyow::Behavior::Verification do
     end
   end
 
-  context "object to verify is defined" do
-    let :params do
+  context "verifiable object is defined" do
+    let(:params) {
       {
         baz: "baz",
         qux: "qux"
       }
-    end
+    }
 
     before do
-      self.class.verifies :params
+      local = self
+      verifiable.class_eval do
+        define_method :params do
+          local.params
+        end
+      end
+
+      verifiable.verifies :params
     end
 
     context "values are passed" do
-      let :values do
+      let(:values) {
         {
           foo: "foo",
           bar: "bar",
         }
-      end
+      }
 
       before do
-        verify values do
+        target.verify values do
           optional :foo
         end
       end
@@ -70,19 +81,19 @@ RSpec.describe Pakyow::Behavior::Verification do
         expect(values).to eq(foo: "foo")
       end
 
-      it "does not verify the object to verify" do
+      it "does not verify the verifiable object" do
         expect(params).to eq(baz: "baz", qux: "qux")
       end
     end
 
     context "values are not passed" do
       before do
-        verify do
+        target.verify do
           optional :baz
         end
       end
 
-      it "verifies the object to verify" do
+      it "verifies the verifiable object" do
         expect(params).to eq(baz: "baz")
       end
     end
@@ -91,20 +102,22 @@ RSpec.describe Pakyow::Behavior::Verification do
   context "verification fails" do
     it "raises an invalid data error" do
       expect {
-        verify bar: "baz" do
+        target.verify bar: "baz" do
           required :foo
         end
       }.to raise_error(Pakyow::InvalidData)
     end
 
     describe "invalid data error" do
-      let :error do
-        verify bar: "baz" do
-          required :foo
+      let(:error) {
+        begin
+          target.verify bar: "baz" do
+            required :foo
+          end
+        rescue => error
+          error
         end
-      rescue => error
-        error
-      end
+      }
 
       it "has the correct message" do
         expect(error.message).to eq("Provided data didn't pass verification")
@@ -120,7 +133,7 @@ RSpec.describe Pakyow::Behavior::Verification do
   context "verification succeeds" do
     it "does not raise an error" do
       expect {
-        verify foo: "foo" do
+        target.verify foo: "foo" do
           required :foo
         end
       }.not_to raise_error